requirements.txt
requests
schedule
ccxt
humanize
loguru
web3
notifiers
apscheduler
python-telegram-bot
ipython
rich
edgex-python-sdk
lighter-sdk@git+https://github.com/Snooowgh/lighter-python.git@main
binance-futures-connector@git+https://github.com/Snooowgh/binance-futures-connector-python.git@main
loguru
ccxt
tenacity
tabulate
pytz
hyperliquid-python-sdk
pybit
python-dotenv
okx==2.1.1
twilio
